Replication added to Reactor
This is done via internal replication RPC events. If actor is not with
replication, it will signal listener that it has no replication implemented.

Note that this replication is just a "placeholder" for future optimizations.
We expect that requesting replicated actor's state will be highly optimized
and a lot of code will be shared between actors.